refactor(cli)!: verb-first command grammar + thin dispatch (phase 1)
Restructure the CLI from a ~2000-line monolith of mixed dispatch mechanisms
(pre-dispatch sys.argv checks, a lazily-built lib subparser, a mutually-exclusive
input group, and early-returning -L/-M/-W/-P flags) into one unconditional
verb-first subparser tree with shared parent parsers.

User-facing impact (BREAKING — clean break, no compat shims):
- Modes are now verbs: `pidcast transcribe <input>`, `pidcast analyze <file>`,
  `pidcast diarize <file>`. Bare `pidcast <url|file>` still shortcuts to transcribe.
- Discovery consolidates under `pidcast list <analyses|models|whisper-models|
  presets|profiles>`, replacing the cryptic -L/-M/-W/-P and --list-chrome-profiles.
- `pidcast paths` -> `pidcast info`.
- `--analyze-existing/--diarize-existing` -> `analyze`/`diarize` verbs;
  `--test-segment` -> `--test` on transcribe.
- Expert whisper-tuning flags (--temperature, --vad-threshold, --whisper-threads,
  --no-speech-thold, --no-fallback, --vad*) are hidden from `transcribe --help`
  for a slimmer surface; still settable and preset-supplied.
- Zero args now prints top-level help instead of an error.

Internals:
- cli.py drops 2014 -> 609 lines; handler bodies relocated verbatim to a new
  commands/ package, plus library/setup/resume domain modules and a duplicate.py.
- resolve_output_dir/_is_legacy_repo_path move to config.py (re-exported from cli
  for compatibility); resume.py reconstructs its namespace via
  build_transcribe_namespace() instead of mutating sys.argv, and filters legacy
  manifest keys so dropped dests can't resurrect.
- New tests/test_cli_grammar.py pins the verb dispatch and bare-input shortcut.

Behavior of the transcription/library workflows is unchanged. 438 passed.
